refactor: Manage 파일 구조 재정리(#47) - Manage 폴더에 Manage.jsx, Manage.css 옮김 - hooks 폴더에 useUserList 옮김 - UseItem 컴포넌트를 list 폴더에서 components 폴더로 옮김
